clippy

+++ crates/cap-test/src/suites/sync.rs
@@ -68,7 +68,7 @@ pub async fn run_suite(hardware: &DiscoveredHardware, duration: u64) -> Result<T
                     dropped: 0,
                     drop_rate_percent: 0.0,
                     effective_fps: metrics.video_frames as f64 / metrics.duration_secs,
-                    target_fps: fps as u32,
+                    target_fps: fps,
                 },
                 latency_ms: LatencyMetrics {
                     avg: 5.0,
